Move QueryStdinForDice to utils
