+++ src/papass/random/dice.py
@@ -1,18 +1,14 @@
 from dataclasses import dataclass
 from typing import Protocol
 
-import click
-
-from papass.utils import rolls_to_value
+from papass.utils import QueryStdinForDice, rolls_to_value
 
 from .base import RngBase
 
 
-# TODO: should appear in docs
 class QueryForDice(Protocol):
     """A callback to be used with ``DiceRng``."""
 
-    # TODO: remove None
     def __call__(self, *, num_sides: int, required_num_rolls: int) -> list[int]:
         """
         :param num_sides: Number of sides of the dice.
@@ -28,59 +24,6 @@ class QueryForDice(Protocol):
 class DiceFrame:
     upper_multiple: int
     required_num_rolls: int
-
-
-# TODO: move this somewhere else
-class QueryStdinForDice(QueryForDice):
-    def __call__(self, *, num_sides: int, required_num_rolls: int) -> list[int]:
-        """Query stdin for desired number of dice rolls.
-
-        Return ``None`` if user gives invalid input.
-        """
-        user_input = input(f"Roll at least {required_num_rolls} dice: ")
-
-        rolls = None
-        while rolls is None:
-            rolls = self._parse_stdin(
-                user_input, num_sides=num_sides, required_num_rolls=required_num_rolls
-            )
-
-        return rolls
-
-    def notify_rejection(self) -> None:
-        """TODO"""
-        click.echo("Rejected. Try again!")
-
-    def _parse_stdin(
-        self, user_input: str, *, num_sides: int, required_num_rolls: int
-    ) -> list[int] | None:
-        """Parse user input as a list of dice rolls.
-
-        Returns ``None`` if input is invalid.
-
-        Example:
-        >>> _parse_stdin("2 3 5", num_sides=6, required_num_rolls=3)
-        [2, 3, 5]
-        """
-        try:
-            rolls = [int(r) for r in user_input.split()]
-        except ValueError:
-            click.echo(
-                "Invalid. Require a space-separated list of integers (like: 1 3 2). Roll again!"
-            )
-            return None
-
-        if len(rolls) < required_num_rolls:
-            click.echo(
-                f"Got only {len(rolls)} rolls, need {required_num_rolls}. Roll again!"
-            )
-            return None
-
-        if not all(1 <= r <= num_sides for r in rolls):
-            click.echo(f"Some rolls are not between 1 and {num_sides}. Roll again!")
-            return None
-
-        return rolls
 
 
 class DiceRng(RngBase):
